Changed Variables in the files at the head from 'int' ---> 'short'

+++ Calculator_Brain/main.c
@@ -1,6 +1,6 @@
 #include "main.h"
 
-int main() {
+short main() {
 littleP();
 simpleC();
   return 0;
